macosx/prefs: always show media key control as enabled
On macOS Sierra and higher, we will use the official API to enable
control of VLC, which will always be enabled as potential conflicts are
mitigated by the OS and it is possible to control more than one
application at the same time, so the option is ignored. On earlier
releases of macOS, SPMediaKeyTap can still be optionally disabled.