contrib: cddb: add dependency to gettext
cddb buildsystem, in particular configure.ac, is using macros from
gettext.m4, which might not be available on non-GNU systems like the
Darwin ones.

When gettext is installed on those systems, the macro file can be
imported using the aclocal's -I parameter or using the ACLOCAL_PATH
environment variable.

When neither is done, even with gettext in the PATH, the configure.ac
won't be able to call AM_ICONV and will trigger the following error.

    configure.ac:130: warning: macro 'AM_ICONV' not found in library
    autoreconf: configure.ac: tracing
    autoreconf: running: libtoolize --copy --force
    libtoolize: putting auxiliary files in '.'.
    libtoolize: copying file './ltmain.sh'
    libtoolize: Consider adding 'AC_CONFIG_MACRO_DIRS([m4])' to configure.ac,
    libtoolize: and rerunning libtoolize and aclocal.
    libtoolize: Consider adding '-I m4' to ACLOCAL_AMFLAGS in Makefile.am.
    configure.ac:130: warning: macro 'AM_ICONV' not found in library
    autoreconf: running: /Users/unidan/Workspace/vlc/extras/tools/build/bin/autoconf --include=/Users/unidan/Workspace/vlc/contrib/aarch64-linux-android/share/aclocal --include=/Users/unidan/Workspace/vlc/extras/tools/build/share/aclocal --force
    configure.ac:130: error: possibly undefined macro: AM_ICONV
          If this token and others are legitimate, please use m4_pattern_allow.
          See the Autoconf documentation.

With the gettext contrib, the gettext.m4 will be installed into the
directory $(PREFIX)/share/aclocal/gettext.m4, and aclocal will be called
with `-I $(PREFIX)/share/aclocal/` which solves the issue here.