avcodec: add a hardware-only lavc AV1 video decoder with higher priority
The priority is higher than dav1d so it can handle hardware decoding even if
dav1d is present.

This version of lavc should only allow decoding if the decoder has a profile
that can be decoded by the hardware, otherwise it should let other decoders try
to do the decoding (dav1d, lavc with current priority, etc).