contrib: lame: Don't use -mtune=native/-march=native
When building a distribution package, tuning specifically for the
system building it is pointless. (If we want to tune specifically
for something else than the compiler's default, we should specify
that and not rely on it being implied from the system doing the
compilation.)

When cross compiling VLC for x86 from a non-x86 system, -mtune=native
and -march=native cause compiler errors (in practice, e.g. for
i686-w64-mingw32 targets built on aarch64, GCC doesn't mind
-mtune=native but errors out for -march=native, while Clang errors
out for -mtune=native too).

This generalizes an existing fix, getting rid of more cases of
-mtune=native/-march=native, applying it for any target, not only
darwin, fixing cross compilation for i686 mingw from aarch64 linux.