vout: improve chrono estimation on start
During a new measurement, the current state of the chrono were always
weighted in the same way, regardless of the number of samples that
contributed in the values. As a consequence, the initial (arbitrary)
value was too impactful.

Instead, until there are enough values (according to the provided
"shift"), weight the average and mean absolute deviation by the number
of contributing samples.

In particular, the first real sample overrides the arbitrary initial
average.