modules: split android_env from android_utils
android_utils depends on GLES2_LIBS and EGL_LIBS, but it isn't forwarded
correctly from the convenience library. In particlar, it contains mostly
video-related code, and some helpers for the android environment state,
which is why the non-video code is using the utils library.

Move the code in a separate convenience library so that it can be
imported without constraints.