swscale: purge obsolete SIMD selection
the swscale cpu feature flags were deprecated in 2012 and removed in 2015
([1]) in favour of auto-detection, hence the version guard used within the
`GetSwsCpuMask()` function.

in vlc v4.0-dev we are purging MMX/MMXEXT and so have a need to remove the
corresponding portions of the function. this alone though would leave the
function and related code only catering to enabling altivec in the one
special case of compiling with an old swscale version. i felt that it was
probably better to just rip it all out.

[1]: https://github.com/FFmpeg/FFmpeg/commit/9d58639e270f7612874681e0ca30fa461e2667b7