video_output: keep track of the next picture when there is one
When we schedule a display of the current frame and it's not late, it will be
scheduled to be displayed at VOUT_REDISPLAY_DELAY. But that will likely be too
late to display the next picture.

So if we can tell when the next frame will need to be displayed, we should use
that information instead.

This partially reverts commit f1bf7ce5b4480a3d38d54c7ae1d1564f0670d83f.