video_output: remove wait misfire hazard margin
The time taken by the prepare is compound of measurements taken in
actual preparation steps. It's made of the time needed to execute the
static filters, the interactive filters, and the time needed for the
display to prepare.

VOUT_MWAIT_TOLERANCE is the margin of error accepted when the scheduler
wakes up the vout_control_Pop deadline late compared to the deadline we
choosed, or wakes up the vlc_clock_Wait late. It's supposed to move the
deadline back wherever relevant to display a frame, to wait a little
less that in the ideal instantaneous or real time case.

It doesn't represent "time spent" so there's no need to account it in
the duration of the prepare as a discriminator for late frames. In
addition, it's hardcoded time not dependant on whether the media or the
system running the video output and accounts for 1/4 of the time budget
for rendering a frame with 60fps.

Regression from a5d85a5f248b12a2ef78041cf2dfa14a6106d66d.