opengl: display: move render code to prepare()
Now that we splitted the vlc_gl_Swap() call from the rendering, there's
no excuse for starting rendering sooner so that the result starts to be
available when executing display() and will almost be ready for
presentation.

is_dirty is there to indicate whether vlc_gl_MakeCurrent failed and thus
the frame could not be rendered, to avoid swapping undefined content.