modules: static.mk: use vlccore.framework on iOS
When compiling the application for iOS, dylibs must be packaged into
frameworks to be installed within the application package. It's easy to
wrap a framework from a dylib after the fact and make the whole
buildsystem believe it has always been a framework.

However, it becomes tricky when enabling bitcode on those platforms,
because bitcode will register the linker commands so that Apple
validation subsystem can recompile it. If we rename into a framework
after the fact, the linker commands registered into the bundle won't be
correct and rebuilding from bitcode will trigger an error.

It's possible to modify the bitcode after the fact to change the bitcode
command, but there's no official objcopy support to modify binary
sections and reinject the new bitcode, and using llvm-objcopy generates
invalid dylibs when modifying a dylib that way.

Make the buildsystem actually believe it's linking against a framework
to have the correct behaviour in the end.