ffmpeg: remove unnecessary header checks
When avcodec is found, it does not really add much value to check for
the avcodec header. avutil is required by it, so checking for that
is not needed either.

For avformat, avcodec is required, therefore checking for
avcodec and avutil headers is not needed there either.

Conditionally including those in avcommon_compat.h does not
seem needed at all, as it won't compile without these headers
anyway.