d3d11: query the driver version using Windows Management Instrumentation
It doesn't work in UWP as that part of the registry is not available via
CoCreateInstanceFromApp(). This might be available via DeviceInformation
from winrt Windows.Devices.Enumeration (not available in wine yet) in the
Properties dictionary, but it would be for UAP/UWP builds only.

Cannot be ported back to 3.0 as this is Vista and above.