macosx: Implement media browsing UI + Local files
Implement the UI for the browse and the streams tabs.
This includes linking to correct icons, correct
NSCollectionViewItem appearance, and correct NSCollectionView
appearance.

Additionally, add local files browsing functionality to
the browse tab.
This integrates NSURLs for local files, retrieved
from NSFileManager, and create VLCInputItems out of them.
This mitigates the need to re-implement the whole functionality
of VLCInputItem for local URLs.