direct3d11: add the texture pointer to fake slices
On NVIDIA and mobile/xbox we let the decoder allocate its own decoder pool
rather than using the one from the vout. But to check that the vout pool has
enough slices for the decoder, we need to have the ID3D11Texture2D used by the
display module.

So far only the real slices had the ID3D11Texture2D. Now we set it on all of
them in case all the good ones are in use, so that the decoder can still probe
what the decoder is using, even if it's not going to use that picture for
anything but probing the number of slices.

And we don't create shader resource view (SRV) for the fake slices.

All fake slices point to slice 0.

Fixes #26009