cdda: Fix handling of GIO CDDA URIs for the whole disk
GIO uses the format "cdda://dev/" for the whole disk, and
"cdda://dev/Track%20N.wav" for specific tracks.  When support for
specific tracks was added to the VLC CDDA module, it broke support for
whole disk URIs with a trailing slash.

This also fixes an infinite loop listing tracks if trying to open track
0, which is invalid. This is because that same open function is used to
open specific tracks when using the "whole disk" mode as well, so this
location parser resetting the track to 0 loops again at discovering all
tracks when tryign to open e.g. "cdda://sr0/Track%200".

Fixes #24767.