http: account for queue in congestion window
The code would credit the congestion window whenever the server had
consumed half of the credits. If the client read data slower than the
server sent it, it would lead to arbitrarily large queues.

To avoid this, only credit the congestion window when the client has
read half the data.

Fixes #26082.