qml: fix search shortcut doesn't gives the searchbox focus
the active focus handler of the local banner was triggered when the search box
got focus through a shortcut and force the focus on another component.

Making the local banner a control fixes the issue. The initial reason of the
change was to be able to check the focus reason in the focus handler, but this
seems unnecessary.