video_output: rename confusing internal variables
The `next_system_pts` has nothing to do with the displayed.next picture. It's
the date of the displayed.current picture in system time.

The `date_next` also has nothing to do with the displayed.next picture. It's
the system date where we should initiate the prepare() for displayed.next.