stats: don't force stats to show right away
If there's a pts, just respect it. Otherwise set the PTS to vlc_tick_now().

The documentation says it should be used like

> vlc -vv http://127.0.0.1:8081 --demux=stats --codec=stats --vout=stats

There's no reason force the "display" (logging) to render right away.

We may not want some of those logs to be dropped because they arrive late. But
we may also want to stat how frames are dropped, which is not possible with
this flag.
To show stats for all frames, even late ones use `--no-drop-late-frames`.