avcapture: use a separate class for ARC
ARC is not supported when storing NSObject in a structure. But it is
well-supported when storing a NSObject as (__bridge_retained void*) into
the p_sys until (__bridge_tranfer) releases it.

It highly simplify the whole handling of ARC objects and avoid using
CFRelease/CFRetain in the code, unifying with other Objective-C modules
like VLCVideoUIView or VLCOpenGLES2VideoView.