opengl: use 3x3 matrix for Android transform
The transform matrix provided by Android is 4x4:
<https://developer.android.com/reference/android/graphics/SurfaceTexture#getTransformMatrix(float%5B%5D)>

However, the third column is never used, since the input vector is in
the form (s, t, 0, 1). Similarly, the third row is never used either,
since only the two first coordinates of the output vector are kept.

Using a 3x3 matrix simplifies the product with other 3x3 matrices.

Note: in theory, a 3x2 matrix would be sufficient, but use a square
matrix for wider compatibility with all OpenGL versions.