macosx: Fix a bug preventing VLC from launching
When the defaults key VideoEffectApplyProfileOnStartup is set to true,
the VLCVideoEffectsWindowController was initialized in a way that caused
the VLCMain's sharedInstance method to be called recursively on launch.

This caused the app to crash, in the VLCMain's init.

The method [VLCVideoEffectsWindowController resetValues] was the source.
In this fix, call this method with dispatch async, to get the
sharedInstance after it is fully intitalized.
This is similar to what is done in VLCAudioEffectsWindowController.

Closes #24746.