doc: QtGL: fix textures being recreated
QOpenGLFramebufferObject::takeTexture is detaching the current texture
and transfering ownership to the caller, and the texture was never
destroyed afterwards. It was leading to a huge memory consumption and
leak when using the sample app.

From the documentation[^1] in Qt5:

    Returns the texture id for the texture attached to this framebuffer
    object. The ownership of the texture is transferred to the caller.

    If the framebuffer object is currently bound, an implicit release()
    will be done. During the next call to bind() a new texture will be
    created.

[1]: https://doc.qt.io/qt-5/qopenglframebufferobject.html#takeTexture

Co-authored-by: Alexandre Janniaux <ajanni@videolabs.io>
For the commit message and cleaning the fix mostly.

Fixes #25867