win32: thread: use the proper DLL to get the system WaitOnAddress
We won't be limited to 32 futeces on platforms that have the API.

The DLL to use is api-ms-win-core-synch-l1-2-0.dll [1].
It could also be in api-ms-win-core-synch-l1-2-1.dll [2].

Maybe it existed as some point in kernel32.dll but it doesn't on my Win10.
Mingw64 has this comment:
; MSDN says it's in Kernel32.dll but it's not.
; Link with libsynchronization.a instead.
; Commented out for compatibility with older
; versions of Windows.
;WaitOnAddress
;WakeByAddressSingle
;WakeByAddressAll

It seems we can use GetModuleHandle() as the DLL is implied by loading kernel32.
This will avoid calling FreeLibrary from DllMain.

In case it doesn't work we fallback to the local implementation of futex.

[1] https://docs.microsoft.com/en-us/windows/win32/api/synchapi/nf-synchapi-waitonaddress
[2] https://docs.microsoft.com/en-us/uwp/win32-and-com/win32-apis#apis-from-api-ms-win-core-synch-l1-2-0dll