opengl: initialize sampler info on creation
A sampler can be initialized in two ways: with or without an interop.
In both cases, it needs the number of input textures and their sizes.

When initialized from vlc_gl_sampler_NewFromInterop(), the number of
textures and their sizes were set immediately.

However, when initialized from vlc_gl_sampler_NewFromTexture2D(), the
number of textures was re-computed indirectly from the chroma, and the
texture sizes were updated on each picture (even if they are constant).

To unify the behavior, always initialize the number of textures and
their sizes on sampler creation. This paves the way to move the interop
out of the sampler.