preparser: fix/simplify synchronisation
* Do not share the notification object for two different events.
* Use binary semaphores instead of boolean and condition variables.
* Use an relaxed atomic variable, get rid of unnecessary lock.

In normal cases, the preparse status would not need to be an atomic
variable, as the preparse semaphore serves as proper barrier. However
in the abort case, the status is written by the "wrong" thread.
Still, we can rely on the semaphore to provide a synchronisation
barrier on the status value.