video_filter: opengl: use correct OpenGL flavour
Use the correct OpenGL flavour (GLES or OpenGL) at runtime by providing
both capabilities, fixing the usage of OpenGL filters through GLES2 on
Linux where both are supported.

The filter only has a single implementation, but will probe both OpenGL
and OpenGL ES2 by default. It can be restricted to a specific flavour by
enforcing a gl module to `any` or any specific implementation. Eg.

    --video-filter=opengl{filter=mock,gles=any}
    --video-filter=opengl{filter=mock,gl=any}
    --video-filter=opengl{filter=mock,gl=any,gles=foo}

OpenGL is favoured when probing since it provides much more
functionality and performance optimization opportunities that its
embedded counterpart. However, when USE_OPENGL_ES2 is defined, it means
that OpenGL headers were not available. In this case, it's highly likely
that OpenGL is **not** available at all and we only probe OpenGL ES
instead.

There are guenuine reasons why it could fail on OpenGL but not on
OpenGL ES, like the presence of an extension like the support for
GL_OES_EGL_image_external or the non-support of anything else but
OpenGL ES with a given version.