win32: do not serialise thread exit
For historical reasons, the TLS cleanup was serialised on a critical
section. Now that there is a read/write lock, there are no reasons to
serialise exit. The thread exit process only reads the list head and
elements, does not modify the list per se; only vlc_threadvar_create()
and vlc_threadvar_delete() add and remove elements respectively.

This allows threads to exit in parallel.