libvlc: ensure that vlc_threadvar_set() is set on a proper thread var
If there's an error, the value we set is not actually set. Maybe an old value
is still in there or the variable is gone or something else is blocking.

We can't use the new value in the future and we don't know if the old value is
still in there.