opengl: fix 2x3 matrix multiplication
To multiply two 2x3 matrices, the right matrix must be expanded to 3x3,
by adding a row [0 0 1], so that the dimensions match (2x3 x 3x3 = 2x3).

The matrix multiplication was erroneous: it behaved as if the matrix was
expanded by an implicit row [1 1 1] instead.

The error has been introduced by
31c20009a02f8b8a30fd4c4f0821ceabba30e30d.