avcodec: video: fix draining
A codec which doesn't set AV_CODEC_CAP_DELAY but has the capability
of multi-threads decoding still needs to be drained by an empty
packet. According to the document, it's safe to pass NULL data to
libavcodec decode function, libavcodec will not pass it along to the
codec unless AV_CODEC_CAP_DELAY is set.