input: use input format for category
The input format is fixed for the lifetime of the decoder thread.
The output format is not always available as the packetiser or the
decoder may delay the determination of their output format until they
start outputting packets.

The downside is the assumption that input and output category are the
same. This assumption is already enforced by an assertion, and relied
heavily upon by the ES output for track selection, so it is hardly a
problem.

Fixes #21975