win32: thread: use the performance counter by default for the date
This should have a better precision (<1us), and we don't have to beg the system
to give us a 5 ms precision. We also don't need extra LoadLibray. And we use
the same code between regular/uwp builds.

Some background on the latency and performance (we are in the Win7 case):
https://docs.microsoft.com/en-us/windows/win32/sysinfo/acquiring-high-resolution-time-stamps