wasm-emscripten: build.sh: use emnm instead of llvm-nm
emnm is a wrapper of llvm-nm and is actually located in the SDK binaries
path added in the PATH when the SDK environment is sourced
(/etc/profile.d/emscripten.sh) or when adding the compiler to the PATH
by sourcing emsdk_env.sh like in the CI configuration.

This avoid defining the EMSDK variable when emscripten is installed on
the system.