opengl: fix orientation matrix initialization
The orientation matrix was not initialized if orientation was
ORIENT_NORMAL.

The regression has been introduced by commit
01db27e13f2bab84d9921694b81ca3743851ee49: since the matrix has been
reduced to 2x3, all the matrix components were set by the macro
MATRIX_SET(), so the prealable initialization to "identity" has been
removed.

However, the switch did not cover the case ORIENT_NORMAL, which had
nothing to do if the matrix was already initialized to "identity".

In practice, this caused a problem only on Android with hardware
acceleration enabled, because interop_android.c is the only interop
which does not apply a vflip to the input format (see commit
1fee4df0606a38d804ddfa0175b5584b7ddc6591).