apple: build.sh: fix compilation for arm64
With latest SDK, arm64 is returned as architecture name for aarch64
machine targets from Apple, which is not recognized by autoconf, and
breaks compilation of multiple contribs.

We still need to indicate that we want an arm64 SDK, but replace arm64
by aarch64 in build triplets until autoconf fixes this.