config: use correct constants
The value returned by config_GetType() is one of the VLC_VAR_*.
The value of param->item.i_type is one of the CONFIG_ITEM_*.

The function config_GetType() both finds the vlc_param from a name and
converts the type.

Since commit 82956e7bc69c0b29059543c19e143561421e4131, the type is not
converted anymore, so it must be compared against CONFIG_ITEM_*
constants directly.

In practice, an internal vflip flag was not correct, so the OpenGL
filter output was vflipped:

    ./vlc --video-filter=opengl --opengl-filter=mock video.mp4