Revert "wasapi: fix passthrough with DTS WAV"
I hate the myself of 2years ago for the lack of explanion on the
original commit. But DTS 14b inside WAV is now played as PCM, using the
same receiver with the same Windows machine than 2 years ago.

Reverting this commit fix it (for 14b and 16b). Maybe a receiver or a
Windows update fixed it ?

This reverts commit 3c68e5ff1ff8da2936da1db4bd9b4e2c4c93c06b.