configure: force -fno-strict-aliasing for C++ modules using COM objects
This avoids potential compiler optimizations when casting to void** when using
IID_PPV_ARGS or WRL.

If the option is not supported we fail the configure as many important C++
modules may not work as expected due to pointer aliasing. In practice this
should never happen as GCC for mingw-w64 and Clang support this correctly.