interop_vaapi: use eglQueryString from GetProcAddress
This allows using the correct implementation for EGL to query the
information for, even when multiple implementation are existing in
different plugins or modules both in the dynamic case or (with partial
linking) the static case.

This effectively makes the interop independant of which EGL
implementation is provided to OpenGL and allows implementation from vgl
callbacks to use vaapi for decoding.

Fix #25256