egl: remove now useless VLC_OPENGL_EXT_*
EGL is now dynamically detected, so there is no point in signalling that
the implementation supports EGL-based extensions, which were not
compatible with libvlc callbacks.

WGL extensions were already made dynamics through previous[^1] similar
commits, and it was only write and never read.

Android was also setting VLC_OPENGL_EXT_EGL for the sake of completness
but didn't actually made use of it. Most of the checks are made through
the availability of the `GL_OES_EGL_image_external` OpenGL extension
which doesn't involve the OpenGL provider API at all.

[^1]: c48c0a6192eb4c0d138861962b3405d4c4ded80b