libplacebo: support HDR output modes
Technically, this just requires forwarding the colorspace and HDR
metadata information to the libplacebo swapchain.

In practice, this does nothing at all on OpenGL, and for Vulkan it only
works with implementations that provide VK_KHR_swapchain_colorspace and
VK_EXT_hdr_metadata, which in practice means "Windows, Wayland and DRM
only", although we don't support DRM or wayland (yet) so it really just
means "Windows only".

We provide only a very short list of valid configurations, because
anything else is severely limited in practice.