avcodec: simplify lavc_UpdateVideoFormat check
lavc_UpdateVideoFormat is returning an error when init_device is NULL.
In addition, a NULL init_device cannot happens since this function
expects NULL as pointer-to-pointer to actually allow (and disable the
loading of) a NULL decoder device.

The check was necessary when c06a7fd2fdc68ce71d330d3c64f4111c4238b0d2
was introduced, but 1798f25add0dae0fc7c3be5e0aca179af0ba9218 fixed the
return value in this case.

This unify the calling of the function with the `,NULL)` version.