opengl: make scale vars independent of libplacebo
The upscaler and downscaler options are valid for all OpenGL vouts.
The fact that they use libplacebo is an implementation detail (they
could support another backend later).

In practice, this also avoids to crash if libplacebo is disabled:
the variables were defined only if HAVE_LIBPLACEBO, but they were read
in all cases.

Fixes #26315