rtp: add pktinfo to convey the M bit
The M bit is part of the common RTP header, but handling is entirely
dependent on the payload format, so we have to pass it somehow.

The pktinfo structure will probably need to be extended to also convey
data from extension headers, e.g. picture orientation or (AV-1) decoding
dependency data.